Stop grouping reports into events Removes all references to `Event` entity
