+++ src/main/resources/db/migration/V1_24__drop_events.sql
@@ -1,0 +1,7 @@
+-- removes events that grouped reports related to one incident
+
+drop index if exists report_event_id_fk_idx;
+alter table report
+  drop column if exists event_id;
+drop table if exists event;
+drop sequence if exists event_sequence;
