+++ src/main/resources/db/seed/R__01_docker_seed.sql
@@ -1,27 +1,11 @@
 do
 $$
     declare
-        v_event_id  UUID;
         v_report_id UUID;
 
     begin
-        select gen_random_uuid() into v_event_id;
-
-        insert into event (
-            id, event_reference,
-            title, description,
-            event_date_and_time, location,
-            created_at, modified_by
-        )
-        values (
-            v_event_id, nextval('event_sequence'),
-            'Test Incident', 'A test incident',
-            now(), 'MDI',
-            now(), 'IR_REPORTING_OFFICER'
-        );
-
         insert into report (
-            id, event_id, report_reference,
+            id, report_reference,
             title, description,
             type, status, source,
             incident_date_and_time, location,
@@ -30,7 +14,7 @@ $$
             modified_by
         )
         values (
-            gen_random_uuid(), v_event_id, nextval('report_sequence'),
+            gen_random_uuid(), nextval('report_sequence'),
             'Assault', 'Details of assualt on staff',
             'ASSAULT', 'DRAFT', 'NOMIS',
             now(), 'MDI',
