+++ src/test/resources/entity-mapping/sample-report-with-details.json
@@ -17,17 +17,6 @@
   "modifiedAt": "2023-12-05T12:34:56",
   "modifiedBy": "USER1",
   "createdInNomis": false,
-  "event": {
-    "eventReference": "11124143",
-    "eventDateAndTime": "2023-12-05T11:34:56",
-    "location": "MDI",
-    "prisonId": "MDI",
-    "title": "An event occurred",
-    "description": "Details of the event",
-    "createdAt": "2023-12-05T12:34:56",
-    "modifiedAt": "2023-12-05T12:34:56",
-    "modifiedBy": "USER1"
-  },
   "questions": [
     {
       "code": "1",
